+++ example/lib/radar_chart/samples/radar_chart_sample1.dart
@@ -120,6 +120,7 @@ class _RadarChartSample1State extends State<RadarChartSample1> {
                 radarBackgroundColor: Colors.transparent,
                 borderData: FlBorderData(show: false),
                 radarBorderData: const BorderSide(color: Colors.transparent),
+                //max: 1000,
                 titlePositionPercentageOffset: 0.2,
                 titleTextStyle:
                     const TextStyle(color: titleColor, fontSize: 14),